+++ .editorconfig
@@ -1,0 +1,6 @@
+[*]
+end_of_line = lf
+
+[*.{ts,tsx,js,jsx}]
+indent_style = space
+indent_size = 2
