+++ README.md
@@ -1,1 +1,10 @@
 # nonogram-player
+
+# Notes
+- introduce unit testing (jest is great! and it's already added); many companies are looking for this. +1 for TDD!!!
+- great use of separation of logic. it's easy to reason about your doc structure
+- front end app developers should know how to work with css compilers like Sass. introduce that here, even for basic css
+- add some basic build automation like CircleCI or Github Actions. They have some basic templates for React applications that run tests and builds for you. Also good to show you know something about this, even in its basic form.
+- in interviews, be able to explain callback functions and function scope. this is VERY important in Javascript, specifically React
+- I added an editor config file just to give mention to it. software engineering is a very collaborative industry, but many people work in many different environments. config files like tsconfig and eslint are intended to enforce patterns scoped to a project. editor config is one more that helps with general formatting across platforms (*nix, Windows)
+- keep your build warnings as clear as possible or unset configured rules; big projects that spit out gobs of warnings that never get addressed can be irritating to work on
