+++ package.json
@@ -4,20 +4,22 @@
   "private": true,
   "homepage": "https://drewrh.github.io/nonogram-player/",
   "dependencies": {
+    "gh-pages": "^3.2.3",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-scripts": "5.0.0",
+    "react-timer-hook": "^3.0.5",
+    "web-vitals": "^2.1.2"
+  },
+  "devDependencies": {
+    "typescript": "^4.5.4",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.0.3",
     "@types/node": "^16.11.14",
     "@types/react": "^17.0.37",
-    "@types/react-dom": "^17.0.11",
-    "gh-pages": "^3.2.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
-    "react-timer-hook": "^3.0.5",
-    "typescript": "^4.5.4",
-    "web-vitals": "^2.1.2"
+    "@types/react-dom": "^17.0.11"
   },
   "scripts": {
     "predeploy": "npm run build",
