+++ src/components/Board/Board.d.ts
@@ -1,0 +1,7 @@
+// specific type definitions can live in definition file within
+// the entity's folder to make them easy to find
+
+interface BoardRowData {
+  newRows: Array<Array<number>>;
+  newRowNums: Array<Array<number>>;
+}
