+++ src/components/components.d.ts
@@ -1,0 +1,20 @@
+interface BoardProps {
+  numRows: number;
+  numColumns: number;
+}
+
+interface Size {
+  width: number | undefined;
+  height: number | undefined;
+}
+
+interface ControlsProps {
+  numRows: number;
+  numColumns: number;
+  changeNumRows: any;
+  changeNumColumns: any;
+}
+
+interface Square {
+  value: number;
+}
